+++ contracts/protocol/libraries/math/TickBitmap.sol
@@ -11,9 +11,9 @@ library TickBitmap {
     /// @param tick The tick for which to compute the position
     /// @return wordPos The key in the mapping containing the word in which the bit is stored
     /// @return bitPos The bit position in the word where the flag is stored
-    function position(int24 tick) private pure returns (int16 wordPos, uint8 bitPos) {
-        wordPos = int16(tick >> 8);
-        bitPos = uint8(tick % 256);
+    function position(int256 tick) private pure returns (int256 wordPos, uint256 bitPos) {
+        wordPos = int256(tick >> 8);
+        bitPos = uint256(tick % 256);
     }
 
     /// @notice Flips the initialized state for a given tick from false to true, or vice versa
@@ -21,12 +21,12 @@ library TickBitmap {
     /// @param tick The tick to flip
     /// @param tickSpacing The spacing between usable ticks == 1
     function flipTick(
-        mapping(int16 => uint256) storage self,
-        int24 tick,
-        int24 tickSpacing
+        mapping(int256 => uint256) storage self,
+        int256 tick,
+        int256 tickSpacing
     ) internal {
         require(tick % tickSpacing == 0); // ensure that the tick is spaced
-        (int16 wordPos, uint8 bitPos) = position(tick / tickSpacing);
+        (int256 wordPos, uint256 bitPos) = position(tick / tickSpacing);
         uint256 mask = 1 << bitPos;
         self[wordPos] ^= mask;
     }
@@ -40,16 +40,16 @@ library TickBitmap {
     /// @return next The next initialized or uninitialized tick up to 256 ticks away from the current tick
     /// @return initialized Whether the next tick is initialized, as the function only searches within up to 256 ticks
     function nextInitializedTickWithinOneWord(
-        mapping(int16 => uint256) storage self,
-        int24 tick,
-        int24 tickSpacing,
+        mapping(int256 => uint256) storage self,
+        int256 tick,
+        int256 tickSpacing,
         bool lte
-    ) internal view returns (int24 next, bool initialized) {
-        int24 compressed = tick / tickSpacing;
+    ) internal view returns (int256 next, bool initialized) {
+        int256 compressed = tick / tickSpacing;
         if (tick < 0 && tick % tickSpacing != 0) compressed--; // round towards negative infinity
 
         if (lte) {
-            (int16 wordPos, uint8 bitPos) = position(compressed);
+            (int256 wordPos, uint256 bitPos) = position(compressed);
             // all the 1s at or to the right of the current bitPos
             uint256 mask = (1 << bitPos) - 1 + (1 << bitPos);
             uint256 masked = self[wordPos] & mask;
@@ -58,11 +58,11 @@ library TickBitmap {
             initialized = masked != 0;
             // overflow/underflow is possible, but prevented externally by limiting both tickSpacing and tick
             next = initialized
-            ? (compressed - int24(bitPos - BitMath.mostSignificantBit(masked))) * tickSpacing
-            : (compressed - int24(bitPos)) * tickSpacing;
+            ? (compressed - int256(bitPos - BitMath.mostSignificantBit(masked))) * tickSpacing
+            : (compressed - int256(bitPos)) * tickSpacing;
         } else {
             // start from the word of the next tick, since the current tick state doesn't matter
-            (int16 wordPos, uint8 bitPos) = position(compressed + 1);
+            (int256 wordPos, uint256 bitPos) = position(compressed + 1);
             // all the 1s at or to the left of the bitPos
             uint256 mask = ~((1 << bitPos) - 1);
             uint256 masked = self[wordPos] & mask;
@@ -71,8 +71,8 @@ library TickBitmap {
             initialized = masked != 0;
             // overflow/underflow is possible, but prevented externally by limiting both tickSpacing and tick
             next = initialized
-            ? (compressed + 1 + int24(BitMath.leastSignificantBit(masked) - bitPos)) * tickSpacing
-            : (compressed + 1 + int24(type(uint8).max - bitPos)) * tickSpacing;
+            ? (compressed + 1 + int256(BitMath.leastSignificantBit(masked) - bitPos)) * tickSpacing
+            : (compressed + 1 + int256(type(uint256).max - bitPos)) * tickSpacing;
         }
     }
 }
